parent ad620cf5a1118c632f4e9ab305e8263490e950c7 author Bingqin Zhou <zhou.bing@husky.neu.edu> 1620934202 -0700 committer Bingqin Zhou <zhou.bing@husky.neu.edu> 1621977830 -0700
DBZ-3506 Remove SchemaProcessor.

DBZ-3506 Nit change.

DBZ-3506 Fix context clean up order.

DBZ-3506 No need to remove all tables in onKeySpaceRemove since this triggers onTableRemove actually.

DBZ-3506 Add fix based on PR 31

[release] Stable parent 1.6.0.Beta1 for release

[maven-release-plugin] prepare release v1.6.0.Beta1

[maven-release-plugin] prepare for next development iteration

[release] New parent 1.6.0-SNAPSHOT for development

DBZ-3506 Address comments.
